requirements.txt
flask
flask-session
langchain
langchain-google-genai
langchain-community
langchain-huggingface
faiss-cpu
python-dotenv
pypdf
sentence-transformers
werkzeug
watchdog
PyPDF2
pytest
pytest-flask
pytest-mock
flask-login
flask-bcrypt
flask-cors
python-json-logger
openai
langchain-openai
langchain-anthropic
cryptography
gunicorn
psycopg2-binary
redis
flask-limiter
tqdm
